Add time-ms, make numbers i64

+++ impls/zig-mal/src/reader.zig
@@ -162,7 +162,7 @@ fn read_atom(allocator: Allocator, reader: *Reader) !*MalType {
             .string = try replaceEscapeSequences(allocator, token[1 .. token.len - 1]),
         } else if (token[0] == ':') MalType{
             .keyword = try allocator.dupe(u8, token[1..]),
-        } else if (std.fmt.parseInt(i32, token, 10)) |int| MalType{
+        } else if (std.fmt.parseInt(i64, token, 10)) |int| MalType{
             .number = int,
         } else |_| MalType{
             .symbol = try allocator.dupe(u8, token),
